chore: reorganize screen commands & messages
